requirements.txt
pyswisseph==2.10.3.2
pytz==2025.1
